requirements.txt
boto3==1.34.108
botocore==1.34.108
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
six==1.16.0
urllib3==2.2.1
